Prefere .bookignore when build too
Reuse watch command experience and process .bookignore (as .gitignore) if prezent.
