+++ src/cmd/watch.rs
@@ -1,3 +1,4 @@
+use crate::utils::ignore::remove_ignored_files;
 use crate::{get_book_dir, open};
 use clap::{arg, App, Arg, ArgMatches};
 use mdbook::errors::Result;
@@ -67,53 +68,6 @@ pub fn execute(args: &ArgMatches) -> Result<()> {
     });
 
     Ok(())
-}
-
-fn remove_ignored_files(book_root: &Path, paths: &[PathBuf]) -> Vec<PathBuf> {
-    if paths.is_empty() {
-        return vec![];
-    }
-
-    match find_gitignore(book_root) {
-        Some(gitignore_path) => {
-            match gitignore::File::new(gitignore_path.as_path()) {
-                Ok(exclusion_checker) => filter_ignored_files(exclusion_checker, paths),
-                Err(_) => {
-                    // We're unable to read the .gitignore file, so we'll silently allow everything.
-                    // Please see discussion: https://github.com/rust-lang/mdBook/pull/1051
-                    paths.iter().map(|path| path.to_path_buf()).collect()
-                }
-            }
-        }
-        None => {
-            // There is no .gitignore file.
-            paths.iter().map(|path| path.to_path_buf()).collect()
-        }
-    }
-}
-
-fn find_gitignore(book_root: &Path) -> Option<PathBuf> {
-    book_root
-        .ancestors()
-        .map(|p| p.join(".gitignore"))
-        .find(|p| p.exists())
-}
-
-fn filter_ignored_files(exclusion_checker: gitignore::File, paths: &[PathBuf]) -> Vec<PathBuf> {
-    paths
-        .iter()
-        .filter(|path| match exclusion_checker.is_excluded(path) {
-            Ok(exclude) => !exclude,
-            Err(error) => {
-                warn!(
-                    "Unable to determine if {:?} is excluded: {:?}. Including it.",
-                    &path, error
-                );
-                true
-            }
-        })
-        .map(|path| path.to_path_buf())
-        .collect()
 }
 
 /// Calls the closure when a book source file is changed, blocking indefinitely.
